Warnings fix for tests. Spotted by cl compiler (#5616)

+++ inference-engine/tests/functional/shared_test_classes/src/subgraph/concat_multi_input.cpp
@@ -70,7 +70,7 @@ void ConcatMultiInput::GenerateConstOnlyModel() {
     ngraph::OutputVector concatInputs;
 
     const int seed = 0;
-    std::mt19937 gen(static_cast<float>(seed));
+    std::mt19937 gen(seed);
 
     auto generateFloatNumbers = [gen](std::size_t vec_len, float min, float max) mutable {
         std::vector<float> res;
